git-svn-id: https://bacula.svn.sourceforge.net/svnroot/bacula/trunk@3094 91ce42f0-d328-0410-95d8-f526ca767f89

+++ bacula/src/lib/base64.c
@@ -157,7 +157,7 @@ bin_to_base64(char *buf, char *bin, int len)
       }
       save = reg;
       reg >>= (rem - 6);
-      buf[j++] = base64_digits[reg & (uint32_t)0x3F];
+      buf[j++] = base64_digits[reg & 0x3F];
       reg = save;
       rem -= 6;
    }
